+++ NEWS.md
@@ -2,11 +2,15 @@
 
 ## Bug fixes
 * Fix bug in `tfrmt_sigdig` so it correctly passes the 'missing' argument to the body_plan (#621, @alanahjonas95).
+* Fix bug in `page_plan` where the max_rows argument was returning an error if a value in the group variable was an empty string (#539, @alanahjonas95).
 
 ## Improvements
 * `footnote_plan()` receives a new argument, `order`, allowing users to specify the order of footnotes (#605, @alanahjonas95).
+* Add markdown processing of stub column labels (#617)
 * Add option to `row_grp_plan()`/`row_grp_structure()` to not add a blank row after the last group (#630, @alanahjonas95).
-   
+
+
+
 # tfrmt 0.3.0
 
 ## Improvements